worker/pyproject.toml
[tool.poetry]
name = "bizkaia-worker"
version = "0.1.0"
description = "Bizkaia Connectivity MVP – Data Pipeline Worker"
authors = ["Bizkaia Connectivity Team"]
packages = [{include = "worker", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.12"
click = ">=8.0,<8.2"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
geoalchemy2 = "^0.15.0"
geopandas = "^0.14"
shapely = "^2.0"
pandas = "^2.2"
pydantic = "^2.7"
pydantic-settings = "^2.3"
structlog = "^24.1"
pyyaml = "^6.0"
httpx = "^0.27"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[tool.poetry.scripts]
bizkaia-worker = "worker.cli:app"

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
